Handle nested errorCode in response
Some 200 responses have `errorCode` as a top-level key. Others nest it
somewhere deeper in the JSON. This is a quick fix to detect errors. A
more complete fix would be to detect which portions of the request
had errors (e.g. which specific recipients we attempted to update failed 
to update).